DRVR/CTRL: Implement triangular (accelerated) move
DRVR:
Add acceleration value applied at each refresh,
which can be set via setAcceleration().
Change accelerate() to setVelocity() which zeroes acceleration.

CTRL:
Extract common calculations into functions.
Implement executeMove(triangular) which accelerates halfway,
then decelerates. This has a different approach from the sleep
of the linear move, busy-waiting for positions to be reached.
As it is not fully robust and the target position may not be
exactly reached with the single deceleration command sent to
the driver, an (arbitrary) minimum speed is enforced at the end.

Co-authored-by: German Germanov <germanov_german@abv.bg>